Divide and update tests
