+++ src/xcode/ENA/ENA/Source/Scenes/QRScanner/_tests_/QRScannerViewModelTests.swift
@@ -39,7 +39,6 @@ class QRScannerViewModelTests: XCTestCase {
 
 	func test_ifValid_PCR_Test_Scanned_then_parsing_is_successful() {
 		let store = MockTestStore()
-		let client = ClientMock()
 		let appConfigurationProvider = CachedAppConfigurationMock()
 		let dscListProvider = MockDSCListProvider()
 		let dccSignatureVerifier = DCCSignatureVerifyingStub()
@@ -47,7 +46,6 @@ class QRScannerViewModelTests: XCTestCase {
 			store: store,
 			dccSignatureVerifier: dccSignatureVerifier,
 			dscListProvider: dscListProvider,
-			client: client,
 			appConfiguration: appConfigurationProvider,
 			cclService: FakeCCLService(),
 			recycleBin: .fake()
@@ -94,7 +92,6 @@ class QRScannerViewModelTests: XCTestCase {
 	
 	func test_ifValid_Antigen_Test_Scanned_then_parsing_is_successful() {
 		let store = MockTestStore()
-		let client = ClientMock()
 		let appConfigurationProvider = CachedAppConfigurationMock()
 		let dscListProvider = MockDSCListProvider()
 		let dccSignatureVerifier = DCCSignatureVerifyingStub()
@@ -102,7 +99,6 @@ class QRScannerViewModelTests: XCTestCase {
 			store: store,
 			dccSignatureVerifier: dccSignatureVerifier,
 			dscListProvider: dscListProvider,
-			client: client,
 			appConfiguration: appConfigurationProvider,
 			cclService: FakeCCLService(),
 			recycleBin: .fake()
@@ -144,7 +140,6 @@ class QRScannerViewModelTests: XCTestCase {
 	
 	func test_ifValid_RapidPCR_Test_Scanned_then_parsing_is_successful() {
 		let store = MockTestStore()
-		let client = ClientMock()
 		let appConfigurationProvider = CachedAppConfigurationMock()
 		let dscListProvider = MockDSCListProvider()
 		let dccSignatureVerifier = DCCSignatureVerifyingStub()
@@ -152,7 +147,6 @@ class QRScannerViewModelTests: XCTestCase {
 			store: store,
 			dccSignatureVerifier: dccSignatureVerifier,
 			dscListProvider: dscListProvider,
-			client: client,
 			appConfiguration: appConfigurationProvider,
 			cclService: FakeCCLService(),
 			recycleBin: .fake()
@@ -193,7 +187,6 @@ class QRScannerViewModelTests: XCTestCase {
 	}
 	func test_ifValid_Event_Scanned_then_parsing_is_successful() {
 		let store = MockTestStore()
-		let client = ClientMock()
 		let appConfigurationProvider = CachedAppConfigurationMock()
 		let dscListProvider = MockDSCListProvider()
 		let dccSignatureVerifier = DCCSignatureVerifyingStub()
@@ -201,7 +194,6 @@ class QRScannerViewModelTests: XCTestCase {
 			store: store,
 			dccSignatureVerifier: dccSignatureVerifier,
 			dscListProvider: dscListProvider,
-			client: client,
 			appConfiguration: appConfigurationProvider,
 			cclService: FakeCCLService(),
 			recycleBin: .fake()
@@ -244,7 +236,6 @@ class QRScannerViewModelTests: XCTestCase {
 	
 	func test_ifValid_Certificate_Scanned_then_parsing_is_successful() {
 		let store = MockTestStore()
-		let client = ClientMock()
 		let appConfigurationProvider = CachedAppConfigurationMock()
 		let dscListProvider = MockDSCListProvider()
 		let dccSignatureVerifier = DCCSignatureVerifyingStub()
@@ -252,7 +243,6 @@ class QRScannerViewModelTests: XCTestCase {
 			store: store,
 			dccSignatureVerifier: dccSignatureVerifier,
 			dscListProvider: dscListProvider,
-			client: client,
 			appConfiguration: appConfigurationProvider,
 			cclService: FakeCCLService(),
 			recycleBin: .fake()
@@ -296,7 +286,6 @@ class QRScannerViewModelTests: XCTestCase {
 	
 	func test_ifInValid_QRCode_Scanned_then_parsing_fails() {
 		let store = MockTestStore()
-		let client = ClientMock()
 		let appConfigurationProvider = CachedAppConfigurationMock()
 		let dscListProvider = MockDSCListProvider()
 		let dccSignatureVerifier = DCCSignatureVerifyingStub()
@@ -304,7 +293,6 @@ class QRScannerViewModelTests: XCTestCase {
 			store: store,
 			dccSignatureVerifier: dccSignatureVerifier,
 			dscListProvider: dscListProvider,
-			client: client,
 			appConfiguration: appConfigurationProvider,
 			cclService: FakeCCLService(),
 			recycleBin: .fake()
@@ -347,7 +335,6 @@ class QRScannerViewModelTests: XCTestCase {
 
 	func testInitialUnsuccessfulScanWithSuccessfulRetry() {
 		let store = MockTestStore()
-		let client = ClientMock()
 		let appConfigurationProvider = CachedAppConfigurationMock()
 		let dscListProvider = MockDSCListProvider()
 		let dccSignatureVerifier = DCCSignatureVerifyingStub()
@@ -355,7 +342,6 @@ class QRScannerViewModelTests: XCTestCase {
 			store: store,
 			dccSignatureVerifier: dccSignatureVerifier,
 			dscListProvider: dscListProvider,
-			client: client,
 			appConfiguration: appConfigurationProvider,
 			cclService: FakeCCLService(),
 			recycleBin: .fake()
