+++ src/xcode/ENA/ENA/Source/Services/CoronaTestService/__tests__/CoronaTestRestorationHandlerTests.swift
@@ -25,7 +25,6 @@ class CoronaTestRestorationHandlerTests: CWATestCase {
 				store: store,
 				dccSignatureVerifier: DCCSignatureVerifyingStub(),
 				dscListProvider: MockDSCListProvider(),
-				client: client,
 				appConfiguration: appConfiguration,
 				cclService: FakeCCLService(),
 				recycleBin: recycleBin
@@ -91,7 +90,6 @@ class CoronaTestRestorationHandlerTests: CWATestCase {
 				store: store,
 				dccSignatureVerifier: DCCSignatureVerifyingStub(),
 				dscListProvider: MockDSCListProvider(),
-				client: client,
 				appConfiguration: appConfiguration,
 				cclService: FakeCCLService(),
 				recycleBin: recycleBin
@@ -136,7 +134,6 @@ class CoronaTestRestorationHandlerTests: CWATestCase {
 				store: store,
 				dccSignatureVerifier: DCCSignatureVerifyingStub(),
 				dscListProvider: MockDSCListProvider(),
-				client: client,
 				appConfiguration: appConfiguration,
 				cclService: FakeCCLService(),
 				recycleBin: recycleBin
