+++ src/xcode/ENA/ENA/Source/Services/HealthCertificate/__tests__/ValidationStateServiceTests.swift
@@ -12,7 +12,6 @@ class TestHealthCertificateService: HealthCertificateService {
 		store: HealthCertificateStoring,
 		dccSignatureVerifier: DCCSignatureVerifying,
 		dscListProvider: DSCListProviding,
-		client: Client,
 		appConfiguration: AppConfigurationProviding,
 		validUntilDates: [Date],
 		expirationDates: [Date]
@@ -21,7 +20,6 @@ class TestHealthCertificateService: HealthCertificateService {
 			store: store,
 			dccSignatureVerifier: dccSignatureVerifier,
 			dscListProvider: dscListProvider,
-			client: client,
 			appConfiguration: appConfiguration,
 			cclService: FakeCCLService(),
 			recycleBin: .fake()
@@ -59,12 +57,10 @@ class ValidationStateServiceTests: XCTestCase {
 
 		let appConfiguration = CachedAppConfigurationMock()
 		let store = MockTestStore()
-		let client = ClientMock()
 		let service = TestHealthCertificateService(
 			store: store,
 			dccSignatureVerifier: DCCSignatureVerifyingStub(),
 			dscListProvider: MockDSCListProvider(),
-			client: client,
 			appConfiguration: appConfiguration,
 			cclService: FakeCCLService(),
 			recycleBin: .fake()
@@ -94,7 +90,6 @@ class ValidationStateServiceTests: XCTestCase {
 			store: store,
 			dccSignatureVerifier: DCCSignatureVerifyingStub(),
 			dscListProvider: dscListProvider,
-			client: ClientMock(),
 			appConfiguration: CachedAppConfigurationMock(),
 			cclService: FakeCCLService(),
 			recycleBin: .fake()
@@ -137,7 +132,6 @@ class ValidationStateServiceTests: XCTestCase {
 			store: MockTestStore(),
 			dccSignatureVerifier: DCCSignatureVerifyingStub(),
 			dscListProvider: MockDSCListProvider(),
-			client: ClientMock(),
 			appConfiguration: CachedAppConfigurationMock(),
 			validUntilDates: dateHelper.orderDates.shuffled(),
 			expirationDates: dateHelper.futureDates.shuffled()
@@ -156,7 +150,6 @@ class ValidationStateServiceTests: XCTestCase {
 			store: MockTestStore(),
 			dccSignatureVerifier: DCCSignatureVerifyingStub(),
 			dscListProvider: MockDSCListProvider(),
-			client: ClientMock(),
 			appConfiguration: CachedAppConfigurationMock(),
 			validUntilDates: dateHelper.orderDates.shuffled(),
 			expirationDates: dateHelper.futureDates.shuffled()
